+++ framework/BaseGameRoom.go
@@ -3,7 +3,7 @@ package framework
 type BaseGameRoom interface {
 GetRoomID() int64
 RegisterConnector(c *BaseSession) error
-FetchConnector(sessionId int64) *BaseSession
+FetchConnector(sessionId int32) *BaseSession
 DeleteConnector(c *BaseSession) error
 BroadcastAll(buff []byte) error
 Unicast(buff []byte, sessionId int64) error
